requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
click==8.1.8
django==4.2.23
djangorestframework==3.16.0
drf-yasg==1.21.7
ebooklib==0.19
gunicorn==21.2.0
joblib==1.5.1
lxml==5.4.0
nltk==3.9.1
regex==2024.11.6
requests==2.31.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.14.0
whitenoise==6.7.0
django-cors-headers==4.3.1
# Image processing and OCR dependencies
Pillow==10.4.0
pytesseract==0.3.10
opencv-python==4.9.0.80
# Language detection
langdetect==1.0.9
# Google Cloud Vision API (optional, for cloud-based OCR)
google-cloud-vision==3.7.1
# Azure AI Vision (optional, for cloud-based OCR) - using compatible version
azure-ai-vision==0.13.0
# AWS Textract (optional, for cloud-based OCR)
boto3==1.34.0
# Spell checking and word validation
pyspellchecker==0.7.2
# Fuzzy string matching for OCR error correction
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
